RowSet: implement IAsyncEnumerable
`RowSet` has had a nasty limitation for a long time: if you wanted
to iterate over all rows in a multi-page result set, you had to do it
synchronously, blocking the calling thread while waiting for the next
page to be fetched. This was not ideal, especially in high-throughput
or UI applications.

Lack of asynchronous iteration was a glaring omission in an otherwise
asynchronous API. It was likely overlooked because the sync API is more
exposed in examples and documentation.

This commit implements `IAsyncEnumerable<Row>` on `RowSet`, allowing
the use of `await foreach` to iterate over all rows in a result set
without blocking the calling thread. This is an additive change,
maintaining backward compatibility with existing synchronous iteration.

Two simple tests is added to verify that async iteration code works.
One just checks the async iteration API with Simulacron.
The other really checks the multipage logic.

Documentation is updated to encourage use of the async API for automatic
paging (`await foreach`).