RowSet: elide GCHandle for columns array
GCHandle brings some (although allegedly small) overhead and complexity.
Since the columns array is only used within the scope of the FFI call,
we can use unsafe code to get a raw pointer to the array instead.
GC cannot invalidate the pointer during the call, because it points
to the stack of the caller, and GC only moves/collects heap objects.