rust: init logger when creating Tokio runtime
Logger must be initialized for the logs to be emitted. As a good
heuristic to initialize it early, we do it when creating the global
Tokio runtime, which happens lazily on the first use of async tasks.
This have a downside that if any logs are emitted before the first
async task is spawned, they will be lost.
To make this more robust, we could consider initializing the logger at
the driver startup, but that would require a new call from C# to Rust,
issued somehow early during the C# driver part.

From now on, the driver will emit logs from Rust code as soon as any async
task is spawned, at a level that is configurable through the env variable
RUST_LOG. For example, if dotnet is run with `RUST_LOG=rust_wrapper=trace`,
all logs from the Rust wrapper will be emitted at trace level, while no
logs from the underlying Rust driver will be emitted.
`RUST_LOG=debug` will emit debug-level logs from both the Rust wrapper
and the underlying Rust driver, and even more complex config is possible:
`RUST_LOG=rust_wrapper=trace,scylla=info`.